Fix auto-import for light R classes in AS 3.3 and 3.4 (KT-27944)
For now R-classes are not indexed (due to performance issues) and because
of that are filtered out in KotlinSourceFilterScope as they looks to
be out of project according to platform indexes.

Also those classes are losing connection with the origin source right
after creation. So normal Kotlin resolve doesn't know how to build a
proper resolver for them (without module or library it's impossible to
build ModuleInfo).

This commit is a workaround for both these problems in auto-import fix:

- We don't wrap resolve scope with KotlinSourceFilterScope
- Classes are resolved with Java resolver build for the desired module

 #KT-27944 Fixed